+++ test/mitmproxy/tools/web/test_master.py
@@ -1,5 +1,4 @@
 import asyncio
-from unittest.mock import MagicMock
 
 import pytest
 
@@ -8,8 +7,11 @@ from mitmproxy.tools.web.master import WebMaster
 
 
 async def test_reuse():
+    async def handler(r, w):
+        pass
+
     server = await asyncio.start_server(
-        MagicMock(), host="127.0.0.1", port=0, reuse_address=False
+        handler, host="127.0.0.1", port=0, reuse_address=False
     )
     port = server.sockets[0].getsockname()[1]
     master = WebMaster(Options(), with_termlog=False)
@@ -18,3 +20,6 @@ async def test_reuse():
     with pytest.raises(OSError, match=f"--set web_port={port + 2}"):
         await master.running()
     server.close()
+    # tornado registers some callbacks,
+    # we want to run them to avoid fatal warnings.
+    await asyncio.sleep(0)